[REFACTOR(player)] 패키지변경
