+++ src/main/kotlin/domain/player/Player.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.player
 
 import domain.card.Card
 import domain.card.CardNumber
