+++ src/test/kotlin/domain/DealerTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.player.Dealer
 import domain.card.Card
 import domain.card.CardNumber
 import domain.card.Shape
